obis: vagrant - fixed git-annex installation for all users

+++ src/vagrant/obis/initialize/install_openbis.sh
@@ -1,13 +1,19 @@
 #!/bin/env bash
 
-pushd . $@ > /dev/null
+if [ ! -L /openbis_installed ]; then
 
-ob_dir=$(find /vagrant/initialize/ -maxdepth 1 -type d -name openBIS-installation-standard-technologies-* | head -n1)
-echo "Installing openBIS from $ob_dir"
-cp /vagrant/config/openbis/console.properties $ob_dir
-chmod a+rw "$ob_dir/console.properties"
+    pushd . $@ > /dev/null
 
-sudo -u openbis cd ~openbis/
-sudo su openbis -c "export ADMIN_PASSWORD=admin && export ETLSERVER_PASSWORD=etlserver && $ob_dir/run-console.sh"
+    ob_dir=$(find /vagrant/initialize/ -maxdepth 1 -type d -name openBIS-installation-standard-technologies-* | head -n1)
+    echo "Installing openBIS from $ob_dir"
+    cp /vagrant/config/openbis/console.properties $ob_dir
+    chmod a+rw "$ob_dir/console.properties"
 
-popd $@ > /dev/null
+    sudo -u openbis cd ~openbis/
+    sudo su openbis -c "export ADMIN_PASSWORD=admin && export ETLSERVER_PASSWORD=etlserver && $ob_dir/run-console.sh"
+
+    sudo touch /openbis_installed
+
+    popd $@ > /dev/null
+
+fi
